move list under lib directory
