+++ app/src/ui/changes/changes-list.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import { CommitMessage } from './commit-message'
 import { ChangedFile } from './changed-file'
-import { List, ClickSource } from '../list'
+import { List, ClickSource } from '../lib/list'
 
 import {
   WorkingDirectoryStatus,
