+++ app/src/ui/history/commit-list.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import { Repository } from '../../models/repository'
 import { Commit } from '../../models/commit'
 import { CommitListItem } from './commit-list-item'
-import { List } from '../list'
+import { List } from '../lib/list'
 import { IGitHubUser } from '../../lib/dispatcher'
 
 const RowHeight = 48
